+++ preprocessing/data_loader.py
@@ -125,35 +125,3 @@ def fit_tokenize_length_threshold(proposition):
         return True
     else:
         return False
-
-
-def prepare_data_for_training(dataset, model_type=None):
-    """Prepare dataset to training. Starts feature extracting.
-    Returns x_data, y_data"""
-    y_data = dataset['label'].to_numpy()
-    numberOfLabels = np.unique(y_data).shape[0]
-    y_data = np.identity(numberOfLabels)[y_data.astype(int).flatten()]
-
-    x_data = dataset.drop(['label', 'argumentationID'], axis=1)
-
-    return x_data, y_data
-	
-def change_labels(dataset, attack=False, bidirect=True):
-        """changes the labels depending on the classification task
-            parameter:
-            dataset: pandas dataframe containing the data
-            attack: if true, the attack label will remain in the dataset, default is false
-            bidirect: allow bidirectional relation instead of one directional
-        """
-
-        if not attack:
-            dataset.loc[dataset.label == 2, 'label'] = 1
-            dataset.loc[dataset.label == -2, 'label'] = -1
-
-        if bidirect:
-            dataset.loc[dataset.label == -1, 'label'] = 1
-        else:
-            dataset.loc[dataset.label == -1, 'label'] = 0
-
-        return dataset
-
